Rename CloudEventRequestToRecordMapper and transform it to a singleton
